tests/acceptance/requirements.txt
Fabric==1.8.3
behave==1.2.5
PyHamcrest==1.8.3
python-keystoneclient==1.3.0
python-glanceclient==0.18.0
oslo.config==1.9.3
oslo.i18n==1.5.0
oslo.serialization==1.4.0
oslo.utils==1.4.0
